add `itemId` to the `delete()` and `get()` method of ExceptionsListApiClient

+++ x-pack/plugins/security_solution/public/management/services/exceptions_list/exceptions_list_api_client.ts
@@ -167,10 +167,14 @@ export class ExceptionsListApiClient {
   }
 
   /**
-   * Returns an item filtered by id
-   * It requires an id in order to get the desired item
+   * Returns an item for the given `itemId` or `id`. Exception List Items have both an `item_id`
+   * and `id`, and at least one of these two is required to be provided.
    */
-  async get(id: string): Promise<ExceptionListItemSchema> {
+  async get(itemId?: string, id?: string): Promise<ExceptionListItemSchema> {
+    if (!itemId && !id) {
+      throw TypeError('either `itemId` or `id` argument must be set');
+    }
+
     await this.ensureListExists;
     return this.http.get<ExceptionListItemSchema>(EXCEPTION_LIST_ITEM_URL, {
       query: {
@@ -204,10 +208,14 @@ export class ExceptionsListApiClient {
   }
 
   /**
-   * It deletes an existing item.
-   * It requires a valid item id.
+   * It deletes an existing item by `itemId` or `id`. Exception List Items have both an `item_id`
+   * and `id`, and at least one of these two is required to be provided.
    */
-  async delete(id: string): Promise<ExceptionListItemSchema> {
+  async delete(itemId?: string, id?: string): Promise<ExceptionListItemSchema> {
+    if (!itemId && !id) {
+      throw TypeError('either `itemId` or `id` argument must be set');
+    }
+
     await this.ensureListExists;
     return this.http.delete<ExceptionListItemSchema>(EXCEPTION_LIST_ITEM_URL, {
       query: {
